+++ modules/edhoc/src/responder.c
@@ -55,7 +55,7 @@ msg1_parse(uint8_t *msg1, uint32_t msg1_len, enum method_type *method,
 {
 	uint32_t i;
 	struct message_1 m;
-	size_t decode_len = 0;
+	uint32_t decode_len = 0;
 
 	TRY_EXPECT(cbor_decode_message_1(msg1, msg1_len, &m, &decode_len),
 		   true);
@@ -147,7 +147,7 @@ static inline enum edhoc_error msg2_encode(const uint8_t *g_y, uint8_t g_y_len,
 					   uint32_t ciphertext_2_len,
 					   uint8_t *msg2, uint32_t *msg2_len)
 {
-	size_t payload_len_out;
+	uint32_t payload_len_out;
 	struct m2 m;
 
 	uint8_t G_Y_CIPHERTEXT_2[g_y_len + ciphertext_2_len];
